(maint) remote total from donut
Having the both the parts and sum of the total made the currently open
PRs donut chart confusing. This just removes the total so it's more
readable. If this CSV is consumed by other tools, we can do this in
another way by hiding the column in the C3 chart definition.