[WIP] first draft of adoptable modules
This is the first draft of a script to generate a list of modules up for
adoption. It will render as an HTML file and a JSON file.

Note that the page is not discoverable without an update to index.html
to link to it. I'll add that once this is reviewed and the cron action
is re-enabled.